+++ tests/testthat/test-call.R
@@ -603,3 +603,11 @@ test_that("is_call_index() works", {
   expect_false(is_call_index(quote(1 + a@b$c[[d]])))
   expect_false(is_call_index(quote((a@b$c[[d]])())))
 })
+
+test_that("call_match() supports `...` in arg list when `dots_expand = FALSE`", {
+  f <- function(x, ...) NULL
+  expect_equal(
+    call_match(quote(f(...)), f, dots_expand = FALSE),
+    quote(f())
+  )
+})